Bug 1143928 - Refactor Deck and Panel classes to reduce code duplication. r=whimboo
