+++ firefox_puppeteer/firefox_puppeteer/ui/about_window/deck.py
@@ -5,6 +5,7 @@
 from marionette_driver import By
 
 from firefox_puppeteer.ui_base_lib import UIBaseLib
+from firefox_puppeteer.ui.panels import BasePanel as Panel
 
 
 class Deck(UIBaseLib):
@@ -131,18 +132,6 @@ class Deck(UIBaseLib):
         :returns: :class:`Panel` instance.
         """
         return self.panels[self.selected_index]
-
-
-class Panel(UIBaseLib):
-
-    def __eq__(self, other):
-        return self.element.get_attribute('id') == other.element.get_attribute('id')
-
-    def __ne__(self, other):
-        return self.element.get_attribute('id') != other.element.get_attribute('id')
-
-    def __str__(self):
-        return self.element.get_attribute('id')
 
 
 class ApplyBillboardPanel(Panel):
