+++ firefox_puppeteer/firefox_puppeteer/ui/pageinfo/deck.py
@@ -5,6 +5,7 @@
 from marionette_driver import By, Wait
 
 from firefox_puppeteer.ui_base_lib import UIBaseLib
+from firefox_puppeteer.ui.panels import BasePanel
 
 
 class Deck(UIBaseLib):
@@ -122,16 +123,7 @@ class Deck(UIBaseLib):
         return panel
 
 
-class Panel(UIBaseLib):
-
-    def __eq__(self, other):
-        return self.element.get_attribute('id') == other.element.get_attribute('id')
-
-    def __ne__(self, other):
-        return self.element.get_attribute('id') != other.element.get_attribute('id')
-
-    def __str__(self):
-        return self.element.get_attribute('id')
+class Panel(BasePanel):
 
     @property
     def tab(self):
